+++ Examples/example_2__json_to_lconf.py
@@ -36,7 +36,7 @@ from LCONF.lconf_structure_classes import (
 from LCONF.main_code import (
    LCONF_DEFAULT,
    lconf_emit,
-   lconf_emit__default_obj,
+   lconf_emit_default_obj,
    lconf_parse_section_extract_by_name,
    lconf_prepare_and_parse_section,
    lconf_prepare_default_obj,
@@ -150,7 +150,7 @@ def main():
    print('  lconf_parse_obj.has_comments: ', lconf_parse_obj.has_comments)
 
    # EXAMPLE 2 a: EMIT DEFAULT OBJ
-   lconf_section_emitted_default_obj_str = lconf_emit__default_obj(
+   lconf_section_emitted_default_obj_str = lconf_emit_default_obj(
       lconf_section__example_2a__template_obj,
       'EMITTED EXAMPLE 2 a',
       onelinelists=LCONF_DEFAULT,
@@ -243,7 +243,7 @@ def main():
    print('  lconf_parse_obj.has_comments: ', lconf_parse_obj.has_comments)
 
    # EXAMPLE 2 b: EMIT DEFAULT OBJ
-   lconf_section_emitted_default_obj_str = lconf_emit__default_obj(
+   lconf_section_emitted_default_obj_str = lconf_emit_default_obj(
       lconf_section__example_2b__template_obj,
       'EMITTED EXAMPLE 2 b',
       onelinelists=LCONF_DEFAULT,